+++ documentation/src/components/index.js
@@ -5,3 +5,5 @@ export * from "./key-features";
 export * from "./perfect-implementation";
 export * from "./get-started";
 export * from "./browser-window";
+export * from "./integrations";
+export * from "./custom-version-dropdown";