+++ packages/mantine/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @pankod/refine-mantine
 
+## 1.0.4
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @pankod/refine-ui-types@0.9.0
+
+## 1.0.3
+
+### Patch Changes
+
+-   Updated dependencies [[`ad99916d6d`](https://github.com/pankod/refine/commit/ad99916d6dbd181b857fd7df7b9619d8cac5e3e0)]:
+    -   @pankod/refine-ui-types@0.8.0
+
 ## 1.0.2
 
 ### Patch Changes